requirements.txt
Django>=4.1.4,<4.2
djangorestframework>=3.13.0,<3.14
djangorestframework-simplejwt>=5.2.2,<5.3
psycopg2>=2.8.6,<2.9
